refactor(query): expose castUpdate as a pure function internally
